+++ CHANGELOG.md
@@ -44,6 +44,7 @@ The following changes have been completed, and will be included in the next tagg
   - added `ExtraPublicData` interface
   - added `NoExtraData` type
   - added `OnError` function signature
+  - added `OnErrorOption` option type
   - added `THROW_THE_ERROR` default error handler
 
 ### Fixes